+++ test/pieces_test.dart
@@ -28,6 +28,8 @@ void main() {
     MoveTest('rbF1', ['e3']),
     MoveTest('vN', ['c2', 'c6', 'e2', 'e6']),
     MoveTest('fslbC', ['a5', 'g5', 'c1']),
+    MoveTest('(4,1)', ['c8', 'e8', 'h3', 'h5']),
+    MoveTest('r(4,3)', ['h1', 'h7']),
   ];
   group('Pieces', () {
     for (CountMovesTest t in countMoveTests) {